authservice/requirements.txt
Flask
Flask-Bcrypt
Flask-JWT-Extended
Flask-Cors
Flask-SQLAlchemy
pyodbc
werkzeug
python-dotenv
